requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
python-dateutil==2.8.2
pytz==2024.1
spotipy==2.23.0
python-dotenv==1.0.1
google-api-python-client==2.108.0 